Refactored main_test.go and added unit-test workflow
